feat: refactor sketches
