+++ src/onpy/features/__init__.py
@@ -1,4 +1,4 @@
 from onpy.features.extrude import Extrude
-from onpy.features.sketch import Sketch
+from onpy.features.sketch.sketch import Sketch
 from onpy.features.planes import Plane, OffsetPlane, DefaultPlane
 from onpy.features.loft import Loft
